feat: remove summit layout from sanity studio
