+++ src/app/(summit-pages)/blocks/speakers.tsx
