+++ src/app/(summit-pages)/register/hubspot-registration-form.scss
